♻️ refactor:  AppDataを移動
Issue #285
